+++ aeordb-lib/spec/engine/v4_first_authority_spec.rs
@@ -50,7 +50,7 @@ fn publisher(algorithm: HashAlgorithm) -> (tempfile::TempDir, Arc<DurabilityCoor
   let directory = tempfile::tempdir().unwrap();
   let path = directory.path().join("first-authority.aeordb");
   let mut file = OpenOptions::new().create_new(true).read(true).write(true).open(path).unwrap();
-  let kv_block_length = initial_block_size() as u64;
+  let kv_block_length = initial_block_size();
   let header = initial_header(algorithm, kv_block_length);
   let slot = encode_database_header_slot(&header).unwrap();
   file.seek(SeekFrom::Start(0)).unwrap();
